add: folder modal-util.js
