+++ js/utils/util.js
@@ -24,6 +24,4 @@ const createRandomIdFromRangeGenerator = (min, max) => {
   };
 };
 
-const isEscapeKey = (evt) => evt.key === 'Escape';
-
-export {getRandomInteger, getRandomArrayElement, createRandomIdFromRangeGenerator, isEscapeKey};
+export {getRandomInteger, getRandomArrayElement, createRandomIdFromRangeGenerator};
